(ENH) begin transition to PyVista for plotting backend
